requirements.txt
# Usage: pip install -r requirements.txt
# Python>=3.6

av>=8.0.3
fastapi>=0.68.0
uvicorn[standard]>=0.15.0

gitpython
ipython
kornia>=0.6.0
matplotlib>=3.2.2
numpy>=1.18.5,<2
opencv-python>=4.1.1
pandas>=1.3.4
pims>=0.5
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
seaborn>=0.11.0
filterpy
tqdm>=4.64.0
torch>=1.10.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.10.0
tensorboard>=2.4.1
Werkzeug>=2.0.3
faiss-cpu==1.7.4
python-multipart
ultralytics>=8.0.0
onnxruntime==1.23.0